fix: bookmark 생성 with constraint
