+++ src/alembic/versions/1a012bd0b76a_init_db.py
@@ -1,0 +1,81 @@
+"""init db
+
+Revision ID: 1a012bd0b76a
+Revises: 
+Create Date: 2021-08-15 19:05:38.719858
+
+"""
+from alembic import op
+import sqlalchemy as sa
+import sqlalchemy_utils
+
+
+# revision identifiers, used by Alembic.
+revision = '1a012bd0b76a'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('tags',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('name', sa.String(length=50), nullable=True),
+    sa.Column('color', sqlalchemy_utils.types.color.ColorType(length=20), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_tags_id'), 'tags', ['id'], unique=False)
+    op.create_index(op.f('ix_tags_name'), 'tags', ['name'], unique=True)
+    op.create_table('users',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('status', sa.Enum('active', 'deleted', 'blocked', name='status'), nullable=True),
+    sa.Column('email', sa.String(length=255), nullable=True),
+    sa.Column('name', sa.String(length=255), nullable=True),
+    sa.Column('profile_img', sa.String(length=1000), nullable=True),
+    sa.Column('sns_type', sa.Enum('apple', 'kakao', 'google', name='sns_type'), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('email')
+    )
+    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
+    op.create_table('bookmarks',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('url', sa.String(length=1000), nullable=True),
+    sa.Column('title', sa.String(length=50), nullable=True),
+    sa.Column('og_img_url', sa.String(length=1000), nullable=True, comment='og 이미지 url'),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('url'),
+    sa.UniqueConstraint('user_id', 'url')
+    )
+    op.create_index(op.f('ix_bookmarks_id'), 'bookmarks', ['id'], unique=False)
+    op.create_index(op.f('ix_bookmarks_user_id'), 'bookmarks', ['user_id'], unique=False)
+    op.create_table('bookmark_tag',
+    sa.Column('bookmark_id', sa.Integer(), nullable=False),
+    sa.Column('tag_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['bookmark_id'], ['bookmarks.id'], ondelete='cascade'),
+    sa.ForeignKeyConstraint(['tag_id'], ['tags.id'], ondelete='cascade'),
+    sa.PrimaryKeyConstraint('bookmark_id', 'tag_id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('bookmark_tag')
+    op.drop_index(op.f('ix_bookmarks_user_id'), table_name='bookmarks')
+    op.drop_index(op.f('ix_bookmarks_id'), table_name='bookmarks')
+    op.drop_table('bookmarks')
+    op.drop_index(op.f('ix_users_id'), table_name='users')
+    op.drop_table('users')
+    op.drop_index(op.f('ix_tags_name'), table_name='tags')
+    op.drop_index(op.f('ix_tags_id'), table_name='tags')
+    op.drop_table('tags')
+    # ### end Alembic commands ###
