--- src/alembic/versions/6b9006d44ab3_initialize.py
@@ -1,54 +1,0 @@
-"""initialize
-
-Revision ID: 6b9006d44ab3
-Revises: 
-Create Date: 2021-08-01 17:12:51.301981
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '6b9006d44ab3'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('status', sa.Enum('active', 'deleted', 'blocked', name='status'), nullable=True),
-    sa.Column('email', sa.String(length=255), nullable=True),
-    sa.Column('name', sa.String(length=255), nullable=True),
-    sa.Column('profile_img', sa.String(length=1000), nullable=True),
-    sa.Column('sns_type', sa.Enum('facebook', 'google', 'kakao', name='sns_type'), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('email')
-    )
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    op.create_table('bookmarks',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('title', sa.String(length=50), nullable=True),
-    sa.Column('url', sa.String(length=1000), nullable=True),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_bookmarks_id'), 'bookmarks', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_bookmarks_id'), table_name='bookmarks')
-    op.drop_table('bookmarks')
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_table('users')
-    # ### end Alembic commands ###
