--- src/alembic/versions/80a9581d6474_add_relation_and_tag_table.py
@@ -1,52 +1,0 @@
-"""add relation and tag table
-
-Revision ID: 80a9581d6474
-Revises: 6b9006d44ab3
-Create Date: 2021-08-08 15:34:06.307702
-
-"""
-from alembic import op
-import sqlalchemy as sa
-import sqlalchemy_utils
-
-
-# revision identifiers, used by Alembic.
-revision = '80a9581d6474'
-down_revision = '6b9006d44ab3'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('tags',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('title', sa.String(length=50), nullable=True),
-    sa.Column('color', sqlalchemy_utils.types.color.ColorType(length=20), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_tags_id'), 'tags', ['id'], unique=False)
-    op.create_table('bookmark_tag',
-    sa.Column('bookmark_id', sa.Integer(), nullable=False),
-    sa.Column('tag_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['bookmark_id'], ['bookmarks.id'], ),
-    sa.ForeignKeyConstraint(['tag_id'], ['tags.id'], ),
-    sa.PrimaryKeyConstraint('bookmark_id', 'tag_id')
-    )
-    op.add_column('bookmarks', sa.Column('favicon_url', sa.String(length=1000), nullable=True))
-    op.add_column('bookmarks', sa.Column('og_url', sa.String(length=1000), nullable=True))
-    op.create_unique_constraint(None, 'bookmarks', ['url'])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'bookmarks', type_='unique')
-    op.drop_column('bookmarks', 'og_url')
-    op.drop_column('bookmarks', 'favicon_url')
-    op.drop_table('bookmark_tag')
-    op.drop_index(op.f('ix_tags_id'), table_name='tags')
-    op.drop_table('tags')
-    # ### end Alembic commands ###
