+++ src/app/crud/bookmark_crud.py
@@ -3,15 +3,16 @@ from sqlalchemy.orm import Session
 from app.models.models.bookmarks import BookmarkCreateRequest, BookmarkUpdateRequest
 from app.services import tag_svc
 from database.schema import Bookmarks
+from fastapi.logger import logger
 
 
 async def create_bookmark(session: Session, user_id, bookmark_in: BookmarkCreateRequest):
     # auto commit = False
     bookmark = Bookmarks.create(session=session, auto_commit=False, user_id=user_id, **bookmark_in.dict())
-    print(bookmark)  # id가 있을까?
-    tags = tag_svc.get_tags_and_create_tags_if_not_existed(session, bookmark.id, bookmark_in.tags)
-    bookmark.tags.append(tags)  # 여기서 북마크에만 처리하면 될까용?
-
+    if bookmark_in.tags:
+        tags = await tag_svc.get_tags_and_create_tags_if_not_existed(session, bookmark_in.tags)
+        logger.info(tags)
+        bookmark.tags.extend(tags)
     session.commit()
     return bookmark
