+++ src/app/crud/tag_crud.py
@@ -1,8 +1,12 @@
 from sqlalchemy.orm import Session
 
+from app.models.models.tags import TagCreateDto
 from database.schema import Tags
 
 
-async def create_tags_by_titles(session: Session, bookmark_id, tag_names):
-    # TODO: bookmark_id 언제 줘야함?
-    return session.bulk_insert_mappings(Tags, [dict(bookmark_id=bookmark_id, name=name) for name in tag_names])
+async def create_tags_by_names(session: Session, tag_names):
+    tags = []
+    for name in tag_names:
+        tag = Tags.create(session=session, auto_commit=False, **TagCreateDto(name=name).dict())
+        tags.append(tag)
+    return tags
